Implement incremental json loading (#389)
## Description
This PR updates the load testing scrambler to load and work on a single
cluster at a time from a very large (~2GB) file of generated data. I
used [`ijson`](https://github.com/ICRAR/ijson) to help with the
incremental json loading, which fortunately allows us to read in each
cluster individually and as a whole rather than in bytes at a time.

To test this, I generated 1.5 million clusters (each with a single
record) using the `gen_seed_data_script.py`. The resulting file was
roughly 2GB (200MB when zipped). I then ran the scrambler on the
unzipped version of the file and was able to sample and generate ~5
million scrambled records in 5 minutes on my local. `python3 -m
tests.load.scripts.scramble_data --file="./test_original_data.json"
--max-seed-records 5000000`

## Related Issues
Closes #387 

## Additional Notes
The next bit of work will be updating the locustfile to reflect the per
cluster randomization and ensuring that we hit the `/seed` endpoint with
only 100 records at a time using `ijson` with a batching function.

<--------------------- REMOVE THE LINES BELOW BEFORE MERGING
--------------------->

## Checklist
Please review and complete the following checklist before submitting
your pull request:

- [x] I have ensured that the pull request is of a manageable size,
allowing it to be reviewed within a single session.
- [x] I have reviewed my changes to ensure they are clear, concise, and
well-documented.
- [x] I have updated the documentation, if applicable.
- [x] I have added or updated test cases to cover my changes, if
applicable.
- [x] I have minimized the number of reviewers to include only those
essential for the review.

## Checklist for Reviewers
Please review and complete the following checklist during the review
process:

- [ ] The code follows best practices and conventions.
- [ ] The changes implement the desired functionality or fix the
reported issue.
- [ ] The tests cover the new changes and pass successfully.
- [ ] Any potential edge cases or error scenarios have been considered.

---------

Co-authored-by: Eric Buckley <eric.buckley@gmail.com>