caching default algo fixture (#414)
## Description
Replacing lru_cache decorator on default_algorithm fixture with a
scope=session for reuse. Updating a handful of tests in test_link.py to
modify a copy of the default_algorithm object, rather than the original,
so we don't break subsequent tests using that same fixture.

## Additional Notes
The latest release of [pytest, 8.4.0, made
changes](https://docs.pytest.org/en/stable/changelog.html#pytest-8-4-0-2025-06-02)
to how fixtures are recognized and loaded. Using an lru_cache decorator
on top of a pytest.fixture, is now causing the harness to not see it.