updates to locustfile (#394)
## Description
This PR updates the locustfile for load testing to:

- Check whether the MPI has been seeded based on user input
- Add a flag for number of records to link to better control the load
testing size
- Loop through the original data (1.5 M records), randomly choose a
record, and send to the link endpoint. Repeat until the number of
records to link has been reached

## Related Issues
Closes #393 

## Additional Notes
If you have better ideas for randomly selecting records from the
original data, I'm all ears, because I'm worried it will make seeding
artificially slow. Right now, the `original_data` file is too large to
load into memory to allow for random indexing so it loops through the
records and randomly decides if they should be linked. It seems unlikely
that we'd test all 1.5M records so the first records would be more
heavily tested. This probably doesn't matter for load testing but I
wanted to point it out.

The other option I considered (but ruled out) was to create the list of
records that would be linked during the seeding process, i.e., when
looping through the `clusters_iter` in the `on_start` section, you could
also randomize whether a cluster gets added to a `record_to_link` list.
However, we'd likely run into the same issue with the list being too
large to hold in memory. We could write the list to file and then read
from that in the `link` task, but it would only remove line 85 `if
random.random() < 0.5`.

Curious to hear other ideas!