Generate class partitions for tuning (#434)
## Description
This PR adds two functions in the mpi_service for creating random
samples of known "true match" and "non-match" data, based on preexisting
labels acquired as part of seeding. The functions use randomization at
the DB query level to improve performance and more quickly generate
possible result sets that can be filtered and returned.

## Related Issues
Closes #429 
Closes #430 

## Additional Notes
This PR also uploads a new file for unit testing the class partitions.
As part of load testing, Marcelle generated a file of 750 thousand
patient clusters. We use a much smaller extract of that here (100
clusters, each of whose patients is actually intended to be similar to
other elements in the cluster) to make sure we're identifying worthwhile
pairs.