Update docker image (#443)
## Description
This PR changes the base image from `[python](python:3.12-slim)` to
`python:3.13-alpine`, which should future proof the image a bit longer
with the upgrade to 3.13 and also switching to alpine should result in
fewer maintenance patches.

## Related Issues
Fixes #440 

## Additional Notes
The Microsoft ODBC driver (msodbcsql18) is not available for Alpine so I
replaced it with `FreeTDS` and `unixODBC`.