Add RMS boundary (#449)
## Description
This PR adds functionality for scoring each of the class-partitioned
pairs of patient records from the DB, and then using those scores to
estimate reasonable bounds for the possible match window (MMT and CMT).
Numerous helper functions got added to tuning to allow us to perform a
modified version of RL's evaluation step (no blocking needed because the
records are already in pairs) with us providing manual overrides to a
number of parameters (namely log-odds).

## Related Issues
Closes #446 

## Additional Notes
As usual, I set up the local server to add this processing step to
hitting the dummy tuning endpoint, and end to end, the entire
computation was ~90 seconds. So that means DB sampling and log-odds
calculating is 60-70s and scoring and sorting the class lists takes
about 20s.