Connect tuning dots (#454)
## Description
This PR ties together the separate tuning pieces--calculation and job
batching--into a single endpoint that we can now use to get results.

## Related Issues
Closes #453 

## Additional Information
Had to make a small change to `LogOdd.validate_feature` because we were
trying to double-invoke `Feature.parse` causing us to try to parse a
feature object rather than a string. The `LogOdd` schema requires
`feature` to be an actual `Feature` type to satisfy mypy, so we need to
pass in an externally constructed `Feature` object. However,
`.validate_feature` is a before validator that operates directly on that
field's input, so we were trying to call `Feature.parse` on something
that was already a `Feature` object to check if it was valid. This was
causing some subtle errors in testing, specifically when trying to
`.split()` the identifier suffix off a feature name. Casting the value
of the Feature field to a string before trying to parse ensures that
mypy and validation are both satisfied.