changing tuning match queries to return iterators (#465)
## Description
Modify the two functions in mpi_service that return true-match and
non-match pairs to be iterators. We have the potential of reducing our
memory footprint by 85% if we don't store the pairs.

## Related Issues
refs #456 

## Additional Notes
This will not resolve #456, this is just the first step of a few to get
us there. There will be follow-up steps to change the prob_calc
functions to accept iterators and then eventually one to change
tuning/base.py::tune to process the data as iterators.

Apologies, but a lot of the changes within are just formatting changes.
My editor automatically runs `ruff format` when I save.

---------

Co-authored-by: Marcelle <53578688+m-goggins@users.noreply.github.com>